refactor: slim time system
